# docs/swagger.yaml
basePath: /api/v1
definitions:
  model.ActionResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        description: 恒为 null
      msg:
        example: 编辑成功
        type: string
    type: object
  model.Diagnose:
    properties:
      buy_num:
        type: integer
      buy_price:
        type: number
      id:
        type: integer
      remark:
        type: string
      report_content:
        type: string
      report_submit_time:
        description: NULL→""
        type: string
      status:
        type: integer
      stock_code:
        type: string
      stock_name:
        type: string
      submit_time:
        type: string
      teacher_name:
        type: string
      user_name:
        type: string
      user_nick_name:
        type: string
    type: object
  model.DiagnoseAuditLog:
    properties:
      operator:
        type: string
      remark:
        type: string
      result:
        type: string
      time:
        type: string
      type:
        type: string
    type: object
  model.DiagnoseAuditReq:
    properties:
      audit_type:
        type: string
      id:
        type: integer
      reject_reason:
        type: string
      result:
        description: pass / reject
        type: string
    type: object
  model.DiagnoseDetail:
    properties:
      audit_logs:
        items:
          $ref: '#/definitions/model.DiagnoseAuditLog'
        type: array
      buy_num:
        type: integer
      buy_price:
        type: number
      id:
        type: integer
      remark:
        type: string
      report_content:
        type: string
      report_submit_time:
        description: NULL→""
        type: string
      status:
        type: integer
      stock_code:
        type: string
      stock_name:
        type: string
      submit_time:
        type: string
      teacher_name:
        type: string
      user_name:
        type: string
      user_nick_name:
        type: string
    type: object
  model.DiagnoseDetailResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.DiagnoseDetail'
      msg:
        example: success
        type: string
    type: object
  model.DiagnoseListResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        properties:
          count:
            example: 15
            type: integer
          list:
            items:
              $ref: '#/definitions/model.Diagnose'
            type: array
        type: object
      msg:
        example: success
        type: string
    type: object
  model.DiagnoseSubmitReportReq:
    properties:
      id:
        type: integer
      report_content:
        description: 富文本 HTML
        type: string
    type: object
  model.GetInfoResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        $ref: '#/definitions/model.UserInfo'
      msg:
        example: success
        type: string
    type: object
  model.LoginReq:
    properties:
      password:
        example: admin123
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  model.LoginResp:
    properties:
      code:
        example: 200
        type: integer
      data:
        description: 恒为 null
      expire:
        description: JWT exp（Unix 秒）
        example: 1724054400
        type: integer
      msg:
        example: 登录成功
        type: string
      passwd_expired:
        description: 密码是否过期（暂恒 false）
        example: false
        type: boolean
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
    type: object
...
  model.UserInfo:
    properties:
      avatar:
        example: ""
        type: string
      introduction:
        example: ""
        type: string
      name:
        example: 系统管理员
        type: string
      permissions:
        example:
        - '*:*:*'
        items:
          type: string
        type: array
      roles:
        example:
        - admin
        items:
          type: string
        type: array
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
info:
  contact: {}
  description: |-
    chatSys（老师管理/绑定业务员/离职转移）+ 诊股记录接口。
    统一响应结构 {code, msg, data}；写操作 msg 为约定中文，查询类为 "success"。
    业务接口需 Bearer token（JWT + Redis 白名单，见 PLAN-auth.md）。
  title: im系统诊股 API
  version: "1.0"
paths:
  /dxsf/diagnose/audit:
    post:
      consumes:
      - application/json
      description: professional 专业审核（状态 2）/ compliance 合规审核（状态 4）；通过 → 3/6，驳回 → 5/4（驳回时
        reject_reason 必填，富文本 HTML）
      parameters:
      - description: 审核请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.DiagnoseAuditReq'
      produces:
      - application/json
      responses:
        "200":
          description: msg 为「审核通过」或「已驳回」，data 恒为 null
          schema:
            $ref: '#/definitions/model.ActionResp'
        "400":
          description: 请求体非法 / audit_type 或 result 非法 / 驳回时原因必填 / 当前状态不允许审核
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 诊股记录不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 审核诊股报告
      tags:
      - 诊股记录
  /dxsf/diagnose/detail:
    get:
      description: 诊股记录全字段 + 审核流程日志（audit_logs）
      parameters:
      - description: 诊股记录 ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiagnoseDetailResp'
        "400":
          description: id 缺失或非整数
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 诊股记录不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 诊股记录详情
      tags:
      - 诊股记录
  /dxsf/diagnose/list:
    get:
      consumes:
      - application/json
      description: 分页查询诊股记录，零值/未传筛选字段不参与过滤；昵称/姓名/股票代码/股票名/老师模糊匹配，ID/买入价/持仓数/状态精确匹配
      parameters:
      - description: 诊股记录 ID（精确）
        in: query
        name: id
        type: integer
      - description: 用户昵称（模糊）
        in: query
        name: user_nick_name
        type: string
      - description: 用户姓名（模糊）
        in: query
        name: user_name
        type: string
      - description: 股票代码（模糊）
        in: query
        name: stock_code
        type: string
      - description: 股票名称（模糊）
        in: query
        name: stock_name
        type: string
      - description: 买入价（精确，DECIMAL 等值匹配）
        in: query
        name: buy_price
        type: number
      - description: 持仓数（精确）
        in: query
        name: buy_num
        type: integer
      - description: 老师姓名（模糊）
        in: query
        name: teacher_name
        type: string
      - description: 状态（精确 1-6：1 待提交 2 待审核 3 已驳回 4 待合规 5 合规驳回 6 已通过）
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: status
        type: integer
      - description: 提交时间起（yyyy-MM-dd，与 submit_end_time 成对生效，闭合区间）
        in: query
        name: submit_begin_time
        type: string
      - description: 提交时间止（yyyy-MM-dd）
        in: query
        name: submit_end_time
        type: string
      - description: 报告提交时间起（yyyy-MM-dd，与 report_end_time 成对生效）
        in: query
        name: report_begin_time
        type: string
      - description: 报告提交时间止（yyyy-MM-dd）
        in: query
        name: report_end_time
        type: string
      - description: 页码（默认 1）
        in: query
        name: page_index
        type: integer
      - description: 页大小（默认 10，上限 100）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.DiagnoseListResp'
        "400":
          description: 数值参数格式错误 / status 越界（非 1-6）
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 诊股记录列表
      tags:
      - 诊股记录
  /dxsf/diagnose/submitReport:
    post:
      consumes:
      - application/json
      description: 状态 1/3/5 可提交（首次编写 / 重新提审），提交后状态统一回落 2；report_content 为富文本 HTML
      parameters:
      - description: 提交报告请求体
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.DiagnoseSubmitReportReq'
      produces:
      - application/json
      responses:
        "200":
          description: msg 固定「提交成功」，data 恒为 null
          schema:
            $ref: '#/definitions/model.ActionResp'
        "400":
          description: 请求体非法 / 报告内容为空 / 当前状态不允许提交
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: 诊股记录不存在
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 提交诊股报告
      tags:
      - 诊股记录
...
  /getinfo:
    get:
      description: 返回角色/昵称等（gyz-admin 动态路由依赖 roles 非空）。需登录态
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.GetInfoResp'
        "401":
          description: 登录已过期，请重新登录
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 当前用户信息
      tags:
      - 鉴权
  /login:
    post:
      consumes:
      - application/json
      description: 账号密码登录，签发 Bearer token（Redis 白名单单设备模式：重新登录即互踢旧设备）。phone_code/uuid
        等多余字段忽略不校验
      parameters:
      - description: 登录凭据
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/model.LoginReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.LoginResp'
      summary: 登录
      tags:
      - 鉴权
  /logout:
    post:
      description: 删除 Redis 白名单使 token 立即失效（幂等）。需登录态
      produces:
      - application/json
      responses:
        "200":
          description: msg 为 退出成功
          schema:
            $ref: '#/definitions/model.ActionResp'
        "401":
          description: 登录已过期，请重新登录
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: 退出登录
      tags:
      - 鉴权
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    description: 值格式：Bearer {token}，登录接口签发
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"